+++ Tests/MapboxSearchUITests/MockWebServer/MockResponse.swift
@@ -1,17 +1,75 @@
 import Foundation
+@_implementationOnly import MapboxCoreSearch
+@testable import MapboxSearch
+import Swifter
+import XCTest
 
-enum MockResponse {
-    enum Endpoint: String {
-        case suggest
-        case retrieve
-        case reverse
-        case multiRetrieve = "retrieve/multi"
-        case categoryCafe = "cafe"
-        case categoryHotel = "hotel"
-        case addressSuggest = "autofill/suggest"
-        case addressRetrieve = "autofill/retrieve"
+protocol MockResponse {
+    /// Resource from test bundle containing response JSON.
+    var filepath: String { get }
+
+    /// HTTP Method used for this request, "GET" or "POST".
+    var httpMethod: HttpServer.HTTPMethod { get }
+
+    /// URL path for this request.
+    var path: String { get }
+
+    /// The API type for this mock.
+    /// Uses Core types to support address autofill.
+    static var coreApiType: CoreSearchEngine.ApiType { get }
+}
+
+// MARK: - Geocoding
+
+enum GeocodingMockResponse: MockResponse {
+    case forwardGeocoding
+    case reverseGeocoding
+    case reverseGeocodingSBS
+
+    var filepath: String {
+        let bundle = Bundle(for: MockWebServer<Self>.self)
+        switch self {
+        case .reverseGeocodingSBS:
+            return bundle.path(forResource: "reverse-geocoding-sbs", ofType: "json")!
+        case .reverseGeocoding:
+            return bundle.path(forResource: "geocoding-reverse-geocoding", ofType: "json")!
+        case .forwardGeocoding:
+            return bundle.path(forResource: "mapbox.places.san.francisco", ofType: "json")!
+        }
     }
 
+    var path: String {
+        var path = "/search/v1"
+        switch self {
+        case .forwardGeocoding:
+            path = "/geocoding/v5/mapbox.places/:query"
+        case .reverseGeocoding:
+            path = "/geocoding/v5/mapbox.places/:location"
+
+        case .reverseGeocodingSBS:
+            path += "/:coordinates"
+        }
+
+        return path
+    }
+
+    var httpMethod: HttpServer.HTTPMethod {
+        switch self {
+        case .forwardGeocoding,
+             .reverseGeocoding,
+             .reverseGeocodingSBS:
+            return .get
+        }
+    }
+
+    static var coreApiType: CoreSearchEngine.ApiType {
+        .geocoding
+    }
+}
+
+// MARK: - SBS
+
+enum SBSMockResponse: MockResponse {
     case suggestEmpty
     case suggestMinsk
     case suggestSanFrancisco
@@ -19,24 +77,19 @@ enum MockResponse {
     case suggestWithCoordinates
     case suggestWithMixedCoordinates
     case suggestCategoryWithCoordinates
-    case suggestAddressSanFrancisco
 
     case retrieveSanFrancisco
     case retrieveMinsk
     case retrieveCategory
     case retrievePoi
     case multiRetrieve
-    case retrieveAddressSanFrancisco
 
     case recursion
-    case forwardGeocoding
-    case reverseGeocoding
-    case reverseGeocodingSBS
     case categoryCafe
     case categoryHotelSearchAlongRoute_JP
 
     var filepath: String {
-        let bundle = Bundle(for: MockWebServer.self)
+        let bundle = Bundle(for: MockWebServer<Self>.self)
         switch self {
         case .suggestEmpty:
             return bundle.path(forResource: "suggestions-empty", ofType: "json")!
@@ -62,22 +115,117 @@ enum MockResponse {
             return bundle.path(forResource: "retrieve-poi", ofType: "json")!
         case .recursion:
             return bundle.path(forResource: "recursion", ofType: "json")!
-        case .reverseGeocodingSBS:
-            return bundle.path(forResource: "reverse-geocoding-sbs", ofType: "json")!
-        case .reverseGeocoding:
-            return bundle.path(forResource: "geocoding-reverse-geocoding", ofType: "json")!
-        case .forwardGeocoding:
-            return bundle.path(forResource: "mapbox.places.san.francisco", ofType: "json")!
         case .multiRetrieve:
             return bundle.path(forResource: "retrieve-multi", ofType: "json")!
         case .categoryCafe:
             return bundle.path(forResource: "category-cafe", ofType: "json")!
         case .categoryHotelSearchAlongRoute_JP:
             return bundle.path(forResource: "category-hotel-search-along-route-jp", ofType: "json")!
-        case .suggestAddressSanFrancisco:
-            return bundle.path(forResource: "address-suggestions-san-francisco", ofType: "json")!
+        }
+    }
+
+    var path: String {
+        var path = "/search/v1"
+
+        switch self {
+        case .suggestMinsk:
+            path += "/suggest/Minsk"
+
+        case .suggestSanFrancisco:
+            path += "/suggest/San Francisco"
+
+        case .suggestEmpty,
+             .suggestCategories,
+             .suggestWithCoordinates,
+             .suggestWithMixedCoordinates,
+             .recursion,
+             .suggestCategoryWithCoordinates:
+            path += "/suggest/:query"
+
+        case .retrieveSanFrancisco,
+             .retrieveCategory,
+             .retrieveMinsk,
+             .retrievePoi:
+            path += "/retrieve"
+
+        case .multiRetrieve:
+            path += "/retrieve/multi"
+
+        case .categoryCafe,
+             .categoryHotelSearchAlongRoute_JP:
+            path += "/category/:category"
+        }
+
+        return path
+    }
+
+    var httpMethod: HttpServer.HTTPMethod {
+        switch self {
+        case .suggestMinsk,
+             .suggestSanFrancisco,
+             .suggestEmpty,
+             .suggestCategories,
+             .suggestWithCoordinates,
+             .suggestWithMixedCoordinates,
+             .suggestCategoryWithCoordinates,
+             .recursion,
+             .categoryCafe,
+             .categoryHotelSearchAlongRoute_JP:
+            return .get
+
+        case .multiRetrieve,
+             .retrieveSanFrancisco,
+             .retrieveCategory,
+             .retrieveMinsk,
+             .retrievePoi:
+            return .post
+        }
+    }
+
+    static var coreApiType: CoreSearchEngine.ApiType {
+        .SBS
+    }
+}
+
+// MARK: - Autofill
+
+enum AutofillMockResponse: MockResponse {
+    case suggestAddressSanFrancisco
+    case retrieveAddressSanFrancisco
+
+    var filepath: String {
+        let bundle = Bundle(for: MockWebServer<Self>.self)
+        switch self {
         case .retrieveAddressSanFrancisco:
             return bundle.path(forResource: "address-retrieve-san-francisco", ofType: "json")!
+        case .suggestAddressSanFrancisco:
+            return bundle.path(forResource: "address-suggestions-san-francisco", ofType: "json")!
         }
+    }
+
+    var path: String {
+        var path = "/search/v1"
+
+        switch self {
+        case .suggestAddressSanFrancisco:
+            path = "/autofill/v1/suggest/:query"
+
+        case .retrieveAddressSanFrancisco:
+            path = "/autofill/v1/retrieve/:action.id"
+        }
+
+        return path
+    }
+
+    var httpMethod: HttpServer.HTTPMethod {
+        switch self {
+        case .suggestAddressSanFrancisco,
+             .retrieveAddressSanFrancisco:
+            return .get
+        }
+    }
+
+    static var coreApiType: CoreSearchEngine.ApiType {
+        .autofill
     }
 }